extract economy to a file
